pyproject.toml
[tool.poetry]
name = "pyafl_qemu_trace"
version = "0.1.3"
description = "A pip-installable distribution of afl-qemu-trace."
authors = ["novafacing <rowanbhart@gmail.com>"]
include = ["pyafl_qemu_trace/binaries/*"]
readme = "README.md"
homepage = "https://github.com/novafacing/pyafl_qemu_trace.git"
repository = "https://github.com/novafacing/pyafl_qemu_trace.git"

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^21.4.0"

[tool.poetry.dev-dependencies]
types-setuptools = "^57.4.14"
black = "^22.3.0"
isort = "^5.10.1"
pylint = "^2.13.7"
mypy = "^0.942"
pytest = "^7.1.2"

[tool.poetry.build]
script = "build.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
